# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Treinão Encontro das Águas — 2ª Edição
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Route {
  id: number;
  name: string;
  distanceKm: number;
  description: string;
  /** @nullable */
  participantLimit?: number | null;
}

export interface Event {
  id: number;
  name: string;
  description: string;
  eventDate: string;
  eventTime: string;
  location: string;
  /** @nullable */
  registrationDeadline?: string | null;
  status: string;
  routes: Route[];
}

export type RegistrationInputSex = typeof RegistrationInputSex[keyof typeof RegistrationInputSex];


export const RegistrationInputSex = {
  masculino: 'masculino',
  feminino: 'feminino',
  outro: 'outro',
} as const;

/**
 * @nullable
 */
export type RegistrationInputShirtSize = typeof RegistrationInputShirtSize[keyof typeof RegistrationInputShirtSize] | null;


export const RegistrationInputShirtSize = {
  PP: 'PP',
  P: 'P',
  M: 'M',
  G: 'G',
  GG: 'GG',
  XG: 'XG',
} as const;

export interface RegistrationInput {
  eventId: number;
  routeId: number;
  /** @minLength 2 */
  fullName: string;
  /**
     * @minLength 11
     * @maxLength 14
     * @pattern ^\d{3}\.?\d{3}\.?\d{3}-?\d{2}$
     */
  cpf: string;
  /** @minLength 8 */
  whatsapp: string;
  /** @nullable */
  email?: string | null;
  birthDate: string;
  sex: RegistrationInputSex;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  shirtSize?: RegistrationInputShirtSize;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactWhatsapp?: string | null;
  acceptedTerms: boolean;
  /** @nullable */
  termsVersion?: string | null;
}

export interface RegistrationConfirmation {
  id: number;
  registrationCode: string;
  fullName: string;
  routeName: string;
  eventName: string;
  eventDate: string;
  eventTime: string;
  location: string;
  createdAt: string;
}

export interface AdminRegistration {
  id: number;
  registrationCode: string;
  fullName: string;
  cpf: string;
  whatsapp: string;
  /** @nullable */
  email?: string | null;
  birthDate: string;
  /** @nullable */
  age?: number | null;
  sex: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  shirtSize?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactWhatsapp?: string | null;
  routeId?: number;
  routeName: string;
  status: string;
  acceptedTerms: boolean;
  /** @nullable */
  acceptedTermsAt?: string | null;
  /** @nullable */
  termsVersion?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  createdAt: string;
}

export interface AdminRegistrationList {
  data: AdminRegistration[];
  total: number;
  page: number;
  pageSize: number;
}

export interface RegistrationUpdate {
  routeId?: number;
  fullName?: string;
  cpf?: string;
  whatsapp?: string;
  /** @nullable */
  email?: string | null;
  birthDate?: string;
  sex?: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  shirtSize?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactWhatsapp?: string | null;
  status?: string;
  /** @nullable */
  internalNotes?: string | null;
}

export interface RouteStats {
  routeName: string;
  distanceKm: number;
  count: number;
}

export interface SexStats {
  sex: string;
  count: number;
}

export interface ShirtStats {
  shirtSize: string;
  count: number;
}

export interface StatusStats {
  status: string;
  count: number;
}

export interface AdminStats {
  total: number;
  byRoute: RouteStats[];
  bySex: SexStats[];
  byShirt: ShirtStats[];
  byStatus: StatusStats[];
  recentRegistrations: AdminRegistration[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListAdminRegistrationsParams = {
search?: string;
route?: string;
sex?: string;
status?: string;
shirtSize?: string;
page?: number;
pageSize?: number;
};