# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Treinão Encontro das Águas — 2ª Edição
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/registrations:
    get:
      operationId: listAdminRegistrations
      tags: [admin]
      summary: List all registrations with filters
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: route
          in: query
          required: false
          schema:
            type: string
        - name: sex
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: shirtSize
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated list of registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRegistrationList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/registrations/{id}:
    get:
      operationId: getAdminRegistration
      tags: [admin]
      summary: Get a single registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRegistration"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateAdminRegistration
      tags: [admin]
      summary: Update a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationUpdate"
      responses:
        "200":
          description: Registration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRegistration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/export:
    get:
      operationId: exportRegistrations
      tags: [admin]
      summary: Export registrations as CSV
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AdminRegistration:
      type: object
      required: [id, registrationCode, fullName, cpf, whatsapp, birthDate, sex, routeName, status, acceptedTerms, createdAt]
      properties:
        id:
          type: integer
        registrationCode:
          type: string
        fullName:
          type: string
        cpf:
          type: string
        whatsapp:
          type: string
        email:
          type: ["string", "null"]
        birthDate:
          type: string
        age:
          type: ["integer", "null"]
        sex:
          type: string
        city:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        team:
          type: ["string", "null"]
        shirtSize:
          type: ["string", "null"]
        medicalNotes:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactWhatsapp:
          type: ["string", "null"]
        routeId:
          type: integer
        routeName:
          type: string
        status:
          type: string
        acceptedTerms:
          type: boolean
        acceptedTermsAt:
          type: ["string", "null"]
        termsVersion:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminRegistrationList:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminRegistration"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    RegistrationUpdate:
      type: object
      properties:
        routeId:
          type: integer
        fullName:
          type: string
        cpf:
          type: string
        whatsapp:
          type: string
        email:
          type: ["string", "null"]
        birthDate:
          type: string
        sex:
          type: string
        city:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        team:
          type: ["string", "null"]
        shirtSize:
          type: ["string", "null"]
        medicalNotes:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactWhatsapp:
          type: ["string", "null"]
        status:
          type: string
        internalNotes:
          type: ["string", "null"]
    RouteStats:
      type: object
      required: [routeName, distanceKm, count]
      properties:
        routeName:
          type: string
        distanceKm:
          type: integer
        count:
          type: integer
    SexStats:
      type: object
      required: [sex, count]
      properties:
        sex:
          type: string
        count:
          type: integer
    ShirtStats:
      type: object
      required: [shirtSize, count]
      properties:
        shirtSize:
          type: string
        count:
          type: integer
    StatusStats:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    AdminStats:
      type: object
      required: [total, byRoute, bySex, byShirt, byStatus, recentRegistrations]
      properties:
        total:
          type: integer
        byRoute:
          type: array
          items:
            $ref: "#/components/schemas/RouteStats"
        bySex:
          type: array
          items:
            $ref: "#/components/schemas/SexStats"
        byShirt:
          type: array
          items:
            $ref: "#/components/schemas/ShirtStats"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusStats"
        recentRegistrations:
          type: array
          items:
            $ref: "#/components/schemas/AdminRegistration"