# lib/api-zod/src/generated/types/adminRegistration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Treinão Encontro das Águas — 2ª Edição
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRegistration {
  id: number;
  registrationCode: string;
  fullName: string;
  cpf: string;
  whatsapp: string;
  /** @nullable */
  email?: string | null;
  birthDate: string;
  /** @nullable */
  age?: number | null;
  sex: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  shirtSize?: string | null;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactWhatsapp?: string | null;
  routeId?: number;
  routeName: string;
  status: string;
  acceptedTerms: boolean;
  /** @nullable */
  acceptedTermsAt?: string | null;
  /** @nullable */
  termsVersion?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  createdAt: string;
}