# lib/api-zod/src/generated/types/routeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Treinão Encontro das Águas — 2ª Edição
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStats {
  routeName: string;
  distanceKm: number;
  count: number;
}